pyproject.toml
[project]
name = "agen-ai-autogen-prj1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.3",
    "chromadb>=0.4.22",
    "fastapi>=0.109.0",
    "gradio>=4.19.2",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.0.2",
    "pandas>=2.3.0",
    "prometheus-client>=0.19.0",
    "psutil>=5.9.8",
    "pyautogen>=0.2.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "pypdf>=3.17.1",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.1",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.7",
    "python-multipart>=0.0.6",
    "uvicorn>=0.27.0",
]
